Refactor https related code
