+++ src/routes/devportalRoute.js
@@ -24,17 +24,10 @@ const devportalController = require('../controllers/devportalController');
 const multer = require('multer');
 const storage = multer.memoryStorage()
 const apiDefinition = multer({ storage: storage })
-const { ensureAuthenticated, validateAuthentication } = require('../middlewares/ensureAuthenticated');
+const { ensureAuthenticated, validateAuthentication, enforceMTLS } = require('../middlewares/ensureAuthenticated');
 const constants = require('../utils/constants');
 
-const enforceMTLS = (req, res, next) => {
-    const clientCert = req.connection.getPeerCertificate();
-    if (!clientCert || !clientCert.subject) {
-        return res.status(403).send('Client certificate required');
-    }
-    return next();
-};
-router.get('/mtls/organizations/:orgId', enforceMTLS, devportalService.getOrganization);
+router.get('/b2b/organizations/:orgId', enforceMTLS, devportalService.getOrganization);
 
 router.post('/organizations', validateAuthentication(constants.SCOPES.ADMIN), adminService.createOrganization);
 router.get('/organizations', validateAuthentication(constants.SCOPES.ADMIN), adminService.getOrganizations);
